BSP-134: using classes extracted into bsp-common-lib
