+++ build.gradle
@@ -117,7 +117,7 @@ configurations {
 }
 
 def versions = [
-        bspCommonLib            : '0.0.19',
+        bspCommonLib            : '0.0.22',
         ccdClient               : '4.6.4',
         commonsBeanUtils        : '1.9.4',
         commonsIo               : '2.5',
