+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/bulkscan/transformation/FinRemBulkScanFormTransformerFactory.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.finrem.caseorchestration.service.bulkscan.transforma
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.bsp.common.error.UnsupportedFormTypeException;
+import uk.gov.hmcts.reform.bsp.common.service.transformation.BulkScanFormTransformerFactory;
 
 import javax.annotation.PostConstruct;
 import java.util.HashMap;
@@ -12,23 +13,15 @@ import static java.lang.String.format;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.service.bulkscan.BulkScanForms.FORM_A;
 
 @Component
-public class BulkScanFormTransformerFactory {
+public class FinRemBulkScanFormTransformerFactory extends BulkScanFormTransformerFactory {
 
     @Autowired
     private FormAToCaseTransformer formAToCaseTransformer;
 
-    private static Map<String, BulkScanFormTransformer> bulkScanFormTransformerMap = new HashMap<>();
-
+    @Override
     @PostConstruct
     public void init() {
+        bulkScanFormTransformerMap = new HashMap<>();
         bulkScanFormTransformerMap.put(FORM_A, formAToCaseTransformer);
-    }
-
-    public BulkScanFormTransformer getTransformer(String formType) {
-        if (!bulkScanFormTransformerMap.containsKey(formType)) {
-            throw new UnsupportedFormTypeException(format("Form type \"%s\" is not supported.", formType));
-        }
-
-        return bulkScanFormTransformerMap.get(formType);
     }
 }
